# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RWCG Backend API
  version: 1.0.0
  description: |
    JSON API served by cmd/apiserver (gin) for two related on-chain projects:

    * **CosmicGame / Cosmic Signature** - a bidding game on Arbitrum. Endpoints under
      `/api/cosmicgame/*` expose bids, rounds, prizes (main prize, raffle ETH,
      Chrono Warrior, Endurance Champion), CST (Cosmic Signature NFT) and
      CosmicToken (ERC-20) data, ETH / ERC-20 / NFT donations, staking
      (CST and RandomWalk wallets), marketing rewards, admin events and
      aggregate statistics.
    * **RandomWalk NFT** - an NFT collection with an on-chain marketplace.
      Endpoints under `/api/randomwalk/*` expose token mints, trading history,
      offers, per-token and market statistics, and the beauty-contest (Elo)
      token ranking.

    Every JSON handler (unless noted otherwise) wraps its payload in a common
    envelope `{"status": 1, "error": "", ...}`; errors are reported as
    `{"status": 0, "error": "<message>"}`. A global per-IP token-bucket rate
    limit (50 rps, burst 100) applies to all routes and returns HTTP 429 when
    exceeded; mutating endpoints add stricter per-route limits.

    Route groups can be disabled per deployment with the environment switches
    ENABLE_ROUTES_COSMICGAME, ENABLE_ROUTES_RANDOMWALK and ENABLE_ROUTES_FAQ
    (all default to enabled). Static NFT asset routes (`/images/**`, `/static/**`)
...
    out of scope.
servers:
  - url: https://api.example
...
components:
  securitySchemes:
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Shared-secret admin key compared against env ADMIN_API_KEY. When the variable is unset the guarded endpoints fail closed with HTTP 503.
    RankingAdminKey:
      type: apiKey
      in: header
      name: X-Ranking-Admin-Key
      description: Shared-secret ranking admin key compared against env RANKING_ADMIN_KEY (falls back to ADMIN_API_KEY). When neither is set the endpoint fails closed with HTTP 503.
...
  responses:
...
    TooManyRequests:
      description: Per-IP rate limit exceeded (global token bucket 50 rps / burst 100; mutating endpoints use stricter per-route limits).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Shared error envelope (status=0 with a human-readable message).
      required: [status, error]
      properties:
        status:
          type: integer
          const: 0
        error:
          type: string
...
    FaqQueryRequest:
      type: object
      required: [question]
      properties:
        question: { type: string }
        session_id:
          type: [string, 'null']
          description: Optional conversation session id; a new session is created when omitted.
    FaqQueryResponse:
      type: object
      properties:
        answer: { type: string }
        sources:
          type: array
          items: { type: string }
        session_id: { type: string }
    FaqReindexResponse:
      type: object
      properties:
        status: { type: string }
        documents_indexed: { type: integer }
    FaqProxyError:
      type: object
      description: Emitted by the Go proxy (HTTP 502) when the Python faq-bot backend is unreachable.
      properties:
        status: { type: integer, const: 0 }
        error: { type: string }
        component: { type: string }
        upstream: { type: string }
paths:
...
  /api/cosmicgame/faq/query:
    post:
      tags: [faq]
      summary: Ask the FAQ bot (proxied)
      description: Proxies to POST /api/query of the Python faq-bot backend. Upstream error responses (400 validation, 410 expired session, 503 subsystem unavailable) are FastAPI-style bodies passed through verbatim.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FaqQueryRequest' }
      responses:
        '200':
          description: Answer with sources and session id.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqQueryResponse' }
        '400':
          description: Empty question (upstream validation error, FastAPI detail body).
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '410':
          description: Session expired (upstream).
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '502':
          description: FAQ backend unreachable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqProxyError' }
        '503':
          description: Upstream subsystem (haystack / codex_mcp) unavailable.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /api/cosmicgame/faq/reindex:
    post:
      tags: [faq]
      summary: Re-index the FAQ knowledge base (proxied)
      description: Proxies to POST /api/reindex of the Python faq-bot backend.
      responses:
        '200':
          description: Re-index finished.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqReindexResponse' }
        '500':
          description: Upstream indexing failure (FastAPI detail body).
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '502':
          description: FAQ backend unreachable.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FaqProxyError' }
        '503':
          description: Upstream retriever not initialized.
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...